[gui] fix non-int input for text width draw

+++ src/pymor/discretizers/builtin/gui/gl.py
@@ -312,7 +312,7 @@ class ColorBarWidget(QOpenGLWidget):
             gl.glVertex(0.5, (bar_height*y + bar_start), y)
         gl.glEnd()
         p.endNativePainting()
-        p.drawText(round((self.width() - self.vmax_width)/2), self.text_ascent, self.vmax_str)
-        p.drawText(round((self.width() - self.vmin_width)/2), self.height() - self.text_height + self.text_ascent,
+        p.drawText(int(round((self.width() - self.vmax_width)/2)), self.text_ascent, self.vmax_str)
+        p.drawText(int(round((self.width() - self.vmin_width)/2)), self.height() - self.text_height + self.text_ascent,
                    self.vmin_str)
         p.end()
